change working dir to action user so files can be deleted
Some data files from a previous run of this action are created with root
as the owner. Because the action runs as a non-root user, such files are
[not able to be deleted](https://github.com/cmu-delphi/forecast-eval/runs/3351290227?check_suite_focus=true)
which is needed for the actions/checkout@v2 step
to install the newest version of the repo.

This seems to be a bug in the actions/checkout@v2 command, perhaps
specific to self-hosted workflows. [Summary of
issue](https://github.com/actions/runner/issues/691) and [suggested
fixes[(https://github.com/actions/checkout/issues/211).