return asof to being empty on init; define startup observe
The UI defines startup selection values globally, for all users
assigned to that Shiny session. Startup values are treated as constants,
so a hard-coded asof date won't update if the Shiny session persists (is
continually used) across a date boundary that would normally change the
default asof selection. Thus, we need to set the starting asof value on
a user-by-user basis so that it is recalculated when a new dashboard
instance is opened. To do that, set the initial asof value back to empty
("") and run an updater observe on dashboard startup only.

In reality, this is practically unimportant; we don't have enough
dashboard users to persist Shiny sessions very long.